refact: calendar 구조 개선
