+++ src/App.tsx
@@ -8,7 +8,7 @@ import { isElectronEnv } from "./types/common";
 const TimelapseWorkspacePage = lazy(
   () => import("./pages/TimelapseWorkspacePage")
 );
-const Calendar = lazy(() => import("./components/Calendar"));
+const Calendar = lazy(() => import("./features/calendar/components/Calendar"));
 const Settings = lazy(() => import("./features/settings/components/Settings"));
 
 // 로딩 중 표시할 컴포넌트
